+++ attribution/src/main/java/com/affise/attribution/modules/store/StoreApi.kt
@@ -5,7 +5,7 @@ import com.affise.attribution.referrer.AffiseReferrerData
 interface StoreApi {
     fun getInstallReferrerData(): AffiseReferrerData?
 
-    fun isInstallReferrerUpdated(): Boolean = false
+    fun isInstallReferrerUpdated(): Boolean
 
     fun startInstallReferrerRetrieve(onFinished: (() -> Unit)? = null)
 }